fixed tensors on different graph bug with input function

+++ data.py
@@ -20,7 +20,7 @@ def csv_to_npy(filename):
 def load_data():
     x = np.load('quizzes.npy')
     y = np.load('solutions.npy')
-    return tf.cast(x, tf.float32), tf.cast(y, tf.float32)
+    return x.astype(float), y.astype(float)
 
 def train_input_fn(x, y, batch_size):
     dataset = tf.data.Dataset.from_tensor_slices((x, y))
